docs: add counter and fetch stories
