+++ starters/lit-mobx-vaadin/.storybook/main.js
@@ -1,3 +1,3 @@
 module.exports = {
-  stories: ['../**/out-tsc/stories/*.stories.{js,md,mdx}'],
+  stories: ['../**/out-tsc/stories/**/*.stories.{js,md,mdx}'],
 };
